+++ src/longident.ml
@@ -46,26 +46,41 @@ let parse_simple s =
   | [] -> assert false
   | s :: l -> unflatten ~init:(Lident s) l
 
+(* find the first matching pair of parentheses *)
+let rec parentheses lpos opened pos len s =
+  if pos >= len then if opened > 0 then Error () else Ok None
+  else
+    match s.[pos] with
+    | '(' ->
+        let lpos = if opened = 0 then pos else lpos in
+        parentheses lpos (opened + 1) (pos + 1) len s
+    | ')' ->
+        let opened = opened - 1 in
+        if opened = 0 then Ok (Some (lpos, pos))
+        else if opened < 0 then Error ()
+        else parentheses lpos opened (pos + 1) len s
+    | _ -> parentheses lpos opened (pos + 1) len s
+
 (* handle ["A.B.(+.+)"] or ["Vec.(.%.()<-)"] *)
 let parse s =
-  let invalid () =
-    invalid_arg (Printf.sprintf "Ppxlib.Longident.parse: %S" s)
+  let invalid variant =
+    invalid_arg (Printf.sprintf "Ppxlib.Longident.parse(%s): %S" variant s)
   in
-  if String.length s < 1 then invalid ();
-  let open_par = String.index_opt s '(' in
-  let close_par = String.index_opt s ')' in
-  match (s.[0], open_par, close_par) with
-  | ('A' .. 'Z' | 'a' .. 'z' | '_'), None, None -> parse_simple s
-  | _, None, None -> Lident s (* This is a raw operator, no module path *)
-  | _, None, _ | _, _, None -> invalid ()
-  | _, Some l, Some r -> (
-      if Int.(r <> String.length s - 1) then invalid ();
+  if String.length s < 1 then invalid "empty string";
+  let par = parentheses (-1) 0 0 (String.length s) s in
+  match (s.[0], par) with
+  | ('A' .. 'Z' | 'a' .. 'z' | '_'), Ok None -> parse_simple s
+  | _, Ok None -> Lident s (* This is a raw operator, no module path *)
+  | _, Error _ -> invalid "unbalanced parenthesis"
+  | _, Ok (Some (l, r)) -> (
+      if Int.(r <> String.length s - 1) then
+        invalid "right parenthesis misplaced";
       let group =
-        if Int.(r = l + 1) then "()"
-        else String.trim (String.sub s ~pos:(l + 1) ~len:(r - l - 1))
+        let inside = String.trim (String.sub s ~pos:(l + 1) ~len:(r - l - 1)) in
+        if String.(inside = "") then "()" else inside
       in
       if Int.(l = 0) then Lident group
-      else if Char.(s.[l - 1] <> '.') then invalid ()
+      else if Char.(s.[l - 1] <> '.') then invalid "application in path"
       else
         let before = String.sub s ~pos:0 ~len:(l - 1) in
         match String.split_on_char before ~sep:'.' with